requirements.txt
torch==1.7.1
torchvision==0.8.2
numpy
scipy
Pillow
matplotlib
tqdm
scikit-image
